Move decorator to utils/decorators.py
